Feat: 기타 Infrastructure 코드
Answer 생성
-Fixed 구현체를 통해서, 채점이 잘 되는지 테스트.
-숫자 야구 게임의 룰상 각 번호는 겹치지 않아야 하므로, 이에 맞게 처리
