--- src/main/java/com/example/BaseballGame/application/FixedGameAnswerGenerator.java
@@ -1,12 +1,0 @@
-package com.example.BaseballGame.application;
-
-import com.example.BaseballGame.domain.GameIdGenerator;
-
-public class FixedGameAnswerGenerator implements GameIdGenerator {
-
-    @Override
-    public Long generateRandomId() {
-        Long id = 123L;
-        return id;
-    }
-}
